requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
groq==0.4.1
boto3==1.34.0
pypdf==3.17.4
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
hypothesis==6.92.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2
